Update set-the-default-browser-using-group-policy.md
FYI: The user may change this setting during the current session only. The default browser will be reset to the policy definition the next time the user logs on. It appears that the policy description is inaccurate, and may mislead the administrator to believe that the policy is only applied once per user.